+++ es6/expressions-ie11.d.ts
@@ -14,6 +14,7 @@ interface ParserOptions {
 		scope: any,
 		context: DXT.ParserContext
 	) => any;
+	postCompile?: (tag: string, meta: any, expr: any) => void;
 }
 
 interface ExpressionParser extends DXT.Parser {